[update] 24266 알고리즘 수업 - 알고리즘의 수행 시간 5 풀이

+++ CodingTest/CodingTest/24265 알고리즘 수업 - 알고리즘의 수행 시간 4.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int n, ret;
+long long n, ret;
 
 int main()
 {
